+++ src/renderer/constants/fonts.ts
@@ -1,5 +1,3 @@
-export const fonts = {
-  robotoRegular: require('~/renderer/resources/fonts/roboto-regular.woff2'),
-  robotoMedium: require('~/renderer/resources/fonts/roboto-medium.woff2'),
-  robotoLight: require('~/renderer/resources/fonts/roboto-light.woff2'),
-};
+export const FONT_ROBOTO_REGULAR = require('~/renderer/resources/fonts/roboto-regular.woff2');
+export const FONT_ROBOTO_MEDIUM = require('~/renderer/resources/fonts/roboto-medium.woff2');
+export const FONT_ROBOTO_LIGHT = require('~/renderer/resources/fonts/roboto-light.woff2');
