+++ src/renderer/views/tabgroup/index.tsx
@@ -2,34 +2,11 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
 import { App } from './components/App';
-import { fonts } from '../../constants';
 import { ipcRenderer } from 'electron';
+import { injectFonts } from '~/renderer/mixins';
 
 ipcRenderer.setMaxListeners(0);
 
-const styleElement = document.createElement('style');
-
-styleElement.textContent = `
-@font-face {
-  font-family: 'Roboto';
-  font-style: normal;
-  font-weight: 400;
-  src: url(${fonts.robotoRegular}) format('woff2');
-}
-@font-face {
-  font-family: 'Roboto';
-  font-style: normal;
-  font-weight: 500;
-  src: url(${fonts.robotoMedium}) format('woff2');
-}
-@font-face {
-  font-family: 'Roboto';
-  font-style: normal;
-  font-weight: 300;
-  src: url(${fonts.robotoLight}) format('woff2');
-}
-`;
-
-document.head.appendChild(styleElement);
+injectFonts();
 
 ReactDOM.render(<App />, document.getElementById('app'));
